# Speed up GetProp Python keyerrors