--- a/Code/GraphMol/Wrap/props.hpp
+++ b/Code/GraphMol/Wrap/props.hpp
@@ -96,25 +96,23 @@ boost::python::dict GetPropsAsDict(const T &obj, bool includePrivate,
         case RDTypeTag::DoubleTag:
           dict[rdvalue.key] = from_rdvalue<double>(rdvalue.val);
           break;
-        case RDTypeTag::StringTag:
-          {
-            auto value = from_rdvalue<std::string>(rdvalue.val);
-            if (autoConvertStrings) {
-              // Auto convert strings to ints and double if possible
-              int ivalue;
-              if (boost::conversion::try_lexical_convert(value, ivalue)) {
-                dict[rdvalue.key] = ivalue;
-                break;
-              }
-              double dvalue;
-              if (boost::conversion::try_lexical_convert(value, dvalue)) {
-                dict[rdvalue.key] = dvalue;
-                break;
-              }
+        case RDTypeTag::StringTag: {
+          auto value = from_rdvalue<std::string>(rdvalue.val);
+          if (autoConvertStrings) {
+            // Auto convert strings to ints and double if possible
+            int ivalue;
+            if (boost::conversion::try_lexical_convert(value, ivalue)) {
+              dict[rdvalue.key] = ivalue;
+              break;
             }
-            dict[rdvalue.key] = value;
+            double dvalue;
+            if (boost::conversion::try_lexical_convert(value, dvalue)) {
+              dict[rdvalue.key] = dvalue;
+              break;
+            }
           }
-          break;
+          dict[rdvalue.key] = value;
+        } break;
         case RDTypeTag::FloatTag:
           dict[rdvalue.key] = from_rdvalue<float>(rdvalue.val);
           break;
@@ -199,64 +197,76 @@ python::object autoConvertString(const RDOb *ob, const std::string &key) {
 
   return python::object();
 }
+
+static PyObject *rawPy(python::object &&pyobj) {
+  Py_INCREF(pyobj.ptr());
+  return pyobj.ptr();
+}
+
+template <class T>
+PyObject *rawPy(T &&thing) {
+  return rawPy(python::object(thing));
+}
+
 template <class RDOb>
-python::object GetPyProp(const RDOb *obj, const std::string &key,
-                         bool autoConvert) {
+PyObject *GetPyProp(const RDOb *obj, const std::string &key, bool autoConvert) {
+  python::object pobj;
   if (!autoConvert) {
-    return python::object(GetProp<RDOb, std::string>(obj, key));
+    std::string res;
+    if (obj->getPropIfPresent(key, res)) {
+      return rawPy(res);
+    } else {
+      PyErr_SetString(PyExc_KeyError, key.c_str());
+      return nullptr;
+    }
   } else {
-    auto &data = obj->getDict().getData();
+    const auto &data = obj->getDict().getData();
     for (auto &rdvalue : data) {
       if (rdvalue.key == key) {
         try {
           const auto tag = rdvalue.val.getTag();
           switch (tag) {
             case RDTypeTag::IntTag:
-              return python::object(from_rdvalue<int>(rdvalue.val));
+              return rawPy(from_rdvalue<int>(rdvalue.val));
 
             case RDTypeTag::DoubleTag:
-              return python::object(from_rdvalue<double>(rdvalue.val));
+              return rawPy(from_rdvalue<double>(rdvalue.val));
 
             case RDTypeTag::StringTag:
               if (autoConvert) {
-                return autoConvertString(obj, rdvalue.key);
+                pobj = autoConvertString(obj, rdvalue.key);
               }
-              return python::object(from_rdvalue<std::string>(rdvalue.val));
-
+              return rawPy(from_rdvalue<std::string>(rdvalue.val));
             case RDTypeTag::FloatTag:
-              return python::object(from_rdvalue<float>(rdvalue.val));
+              return rawPy(from_rdvalue<float>(rdvalue.val));
               break;
             case RDTypeTag::BoolTag:
-              return python::object(from_rdvalue<bool>(rdvalue.val));
+              return rawPy(from_rdvalue<bool>(rdvalue.val));
               break;
             case RDTypeTag::UnsignedIntTag:
-              return python::object(from_rdvalue<unsigned int>(rdvalue.val));
+              return rawPy(from_rdvalue<unsigned int>(rdvalue.val));
               break;
             case RDTypeTag::AnyTag:
               // we skip these for now
               break;
             case RDTypeTag::VecDoubleTag:
-              return python::object(
-                  from_rdvalue<std::vector<double>>(rdvalue.val));
+              return rawPy(from_rdvalue<std::vector<double>>(rdvalue.val));
               break;
             case RDTypeTag::VecFloatTag:
-              return python::object(
-                  from_rdvalue<std::vector<float>>(rdvalue.val));
+              return rawPy(from_rdvalue<std::vector<float>>(rdvalue.val));
               break;
             case RDTypeTag::VecIntTag:
-              return python::object(
-                  from_rdvalue<std::vector<int>>(rdvalue.val));
+              return rawPy(from_rdvalue<std::vector<int>>(rdvalue.val));
               break;
             case RDTypeTag::VecUnsignedIntTag:
-              return python::object(
+              return rawPy(
                   from_rdvalue<std::vector<unsigned int>>(rdvalue.val));
               break;
             case RDTypeTag::VecStringTag:
-              return python::object(
-                  from_rdvalue<std::vector<std::string>>(rdvalue.val));
+              return rawPy(from_rdvalue<std::vector<std::string>>(rdvalue.val));
               break;
             case RDTypeTag::EmptyTag:
-              return boost::python::object();
+              return Py_None;
               break;
             default:
               std::string message =
@@ -264,7 +274,7 @@ python::object GetPyProp(const RDOb *obj, const std::string &key,
                       "Unhandled property type encountered for property: ") +
                   rdvalue.key;
               UNDER_CONSTRUCTION(message.c_str());
-              return boost::python::object();
+              return Py_None;
           }
         } catch (std::bad_any_cast &) {
           // C++ datatypes can really be anything, this just captures
@@ -273,14 +283,33 @@ python::object GetPyProp(const RDOb *obj, const std::string &key,
               std::string("Unhandled type conversion occured for property: ") +
               rdvalue.key;
           UNDER_CONSTRUCTION(message.c_str());
-          return boost::python::object();
+          return Py_None;
         }
       }
     }
   }
   PyErr_SetString(PyExc_KeyError, key.c_str());
-  throw python::error_already_set();
+  return nullptr;
 }
+
+// Return policy for functions that directly return a PyObject* and
+// are fully responsible for setting the Python error state.
+struct return_pyobject_passthrough {
+  template <class T>
+  struct apply {
+    struct type {
+      static bool convertible() { return true; }
+
+      PyObject *operator()(PyObject *inner) const { return inner; }
+#ifndef BOOST_PYTHON_NO_PY_SIGNATURES
+      PyTypeObject const *get_pytype() const {
+        return boost::python::converter::expected_pytype_for_arg<
+            T>::get_pytype();
+      }
+#endif
+    };
+  };
+};
 
 template <class RDOb>
 int MolHasProp(const RDOb &mol, const std::string &key) {