#### What does this implement/fix? Explain your changes.

A common pattern _in Python_ for checking for the presence or absence of a key is:

```
    try:
       return mol.GetProp('mykey')
    except KeyError:
       return None
```

Shockingly, this is really slow with boost python objects! I was recently profiling a workflow and 90% of the time or more was spent in failed GetProp calls (mostly on bonds, some on atoms or mols).

I sped up the workflow by protecting the calls using HasProp. But I think this is a silly trap we've set for our users.

The problem comes because boost::python uses a C++ exception to indicate that there is already a Python exception set. In C++, exceptions are slow - they require unrolling a stack. In Python, exceptions are about the same speed as any other control flow!

This commit speeds up GetProp failures by circumventing the boost throw_exception_already_set() mechanism. I based `return_pyobject_passthrough` on `return_none` in boost::python's return_arg.hpp https://github.com/boostorg/python/blob/3e7be69e1e405e1d5ddd232c69c024ee441592c5/include/boost/python/return_arg.hpp#L36. It was really annoying to figure out the right incantation here.

In my testing, this speeds up failed GetProp substantially:

* Factor of 1000x on Mac
* Factor of 40x on Linux

For 100,000 Mol.GetProp calls, before this commit:

|   |  Mac  | Linux |
| - | ----- | ----- |
| Present prop | 0.03s | 0.02s |
| Missing prop | 30s | 0.82s |

(after this commit missing and present properties aer about the same duration)