@bp-kelley 2025-03-22
@d-b-w I think this is great and along the same lines I've always wanted to add the more pythonic

```
foo = atom.GetProp("bar", None)
```

And this might put me over the edge if you all think it's a good idea.