@d-b-w 2025-04-04
I've updated this with a fix to the typed GetXProp calls (e.g. GetIntProp).

The test failures don't look related (to me)